refactor: merge Reader and Writer interface and rename package
